+++ app/src/main/java/com/wepli/app/navigation/NavHost.kt
@@ -21,12 +21,12 @@ import com.wepli.app.navigation.extensions.navigateToBack
 import com.wepli.community.navigation.communityWriteGraph
 import com.wepli.community.navigation.navigateToBackAndPostRefresh
 import com.wepli.community.navigation.navigateToCommunityWrite
-import com.wepli.feature.namecard.detail.navigation.photoCardDetailGraph
-import com.wepli.feature.namecard.detail.navigation.navigateToPhotoCardDetail
-import com.wepli.feature.namecard.main.navigation.photoCardMainGraph
-import com.wepli.feature.namecard.main.navigation.navigateToPhotoCardMain
-import com.wepli.feature.namecard.result.navigation.photoCardResultGraph
-import com.wepli.feature.namecard.result.navigation.navigateToPhotoCardResult
+import com.wepli.feature.photocard.detail.navigation.photoCardDetailGraph
+import com.wepli.feature.photocard.detail.navigation.navigateToPhotoCardDetail
+import com.wepli.feature.photocard.main.navigation.photoCardMainGraph
+import com.wepli.feature.photocard.main.navigation.navigateToPhotoCardMain
+import com.wepli.feature.photocard.result.navigation.photoCardResultGraph
+import com.wepli.feature.photocard.result.navigation.navigateToPhotoCardResult
 import com.wepli.playlist.navigation.navigateToPlaylistDetail
 import com.wepli.playlist.navigation.playlistDetailGraph
 import com.wepli.search.navigation.SearchScreenMode
