--- feature/namecard/src/main/java/com/wepli/feature/namecard/main/mvi/NameCardMainMvi.kt
@@ -1,20 +1,0 @@
-package com.wepli.feature.namecard.main.mvi
-
-import base.Intent
-import base.SideEffect
-import base.UiState
-import com.wepli.shared.feature.uimodel.namecard.PhotoCardUiData
-import com.wepli.shared.feature.uimodel.user.UserUiData
-
-data class NameCardMainUiState(
-    val user: UserUiData = UserUiData(),
-    val nameCardInfo: PhotoCardUiData = PhotoCardUiData(),
-) : UiState
-
-interface NameCardMainEffect : SideEffect {
-
-}
-
-interface NameCardMainIntent : Intent {
-
-}
