+++ feature/namecard/src/main/java/com/wepli/feature/photocard/detail/chapter/PhotoCardChapter3Screen.kt
@@ -1,0 +1,97 @@
+package com.wepli.feature.photocard.detail.chapter
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import button.WepliBasicButton
+import button.WepliButtonStyle
+import com.wepli.feature.photocard.detail.mvi.PhotoCardDetailIntent
+import com.wepli.feature.photocard.detail.mvi.PhotoCardDetailUiState
+import textfield.WepliTextField
+import textfield.WepliTextFieldType
+import theme.WepliTheme
+
+@Preview(
+    showBackground = true,
+    backgroundColor = 0xFF000000
+)
+@Composable
+fun PhotoCardChapter3ScreenPreview() {
+    PhotoCardChapter3Screen(
+        state = PhotoCardDetailUiState(),
+        sendAction = {},
+        navOnNextPage = {},
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(top = 56.dp, bottom = 20.dp)
+    )
+}
+
+@Composable
+fun PhotoCardChapter3Screen(
+    state: PhotoCardDetailUiState,
+    sendAction: (PhotoCardDetailIntent) -> Unit,
+    navOnNextPage: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val instagramIdState = state.instagramId
+    val isEnabled by remember(instagramIdState.text) {
+        derivedStateOf { instagramIdState.text.isNotEmpty() && instagramIdState.isLengthExceed.not() }
+    }
+
+    Column(modifier = modifier.padding(horizontal = 24.dp)) {
+        Text(
+            text = "Instagram 계정을\n입력해주세요",
+            style = WepliTheme.typo.title1,
+            color = WepliTheme.color.gray900
+        )
+        Spacer(modifier = Modifier.padding(top = 12.dp))
+        Text(
+            text = "명함에 표시될 SNS 계정을 입력해주세요\n계정 입력을 원하지 않으면 건너뛸 수 있어요",
+            style = WepliTheme.typo.body4,
+            color = WepliTheme.color.gray500
+        )
+
+        Spacer(modifier = Modifier.height(32.dp))
+        WepliTextField(
+            value = "",
+            onValueChanged = { newValue ->
+                sendAction(PhotoCardDetailIntent.OnChangedInstagramId(newValue))
+            },
+            singleLine = true,
+            placeholder = "계정을 입력해주세요",
+            type = WepliTextFieldType.Normal,
+        )
+
+        Spacer(modifier = Modifier.weight(1f))
+        WepliBasicButton(
+            title = "입력완료",
+            isEnabled = isEnabled,
+            onClick = { navOnNextPage() },
+            modifier = Modifier
+                .align(Alignment.CenterHorizontally),
+            buttonStyle = WepliButtonStyle.Basic,
+        )
+        Spacer(modifier = Modifier.height(8.dp))
+        WepliBasicButton(
+            title = "건너뛰기",
+            isEnabled = true,
+            onClick = { navOnNextPage() },
+            modifier = Modifier
+                .padding(bottom = 20.dp)
+                .align(Alignment.CenterHorizontally),
+            buttonStyle = WepliButtonStyle.Transparent,
+        )
+    }
+}
