+++ feature/namecard/src/main/java/com/wepli/feature/photocard/result/NameCardResultViewModel.kt
@@ -1,0 +1,39 @@
+package com.wepli.feature.photocard.result
+
+import base.BaseMviViewModel
+import com.wepli.feature.photocard.result.mvi.NameCardResultEffect
+import com.wepli.feature.photocard.result.mvi.NameCardResultIntent
+import com.wepli.feature.photocard.result.mvi.NameCardResultUiState
+import com.wepli.shared.feature.uimodel.user.UserUiData
+import dagger.hilt.android.lifecycle.HiltViewModel
+import repository.user.UserRepository
+import javax.inject.Inject
+
+@HiltViewModel
+class NameCardResultViewModel @Inject constructor(
+    private val userRepository: UserRepository
+) : BaseMviViewModel<NameCardResultUiState, NameCardResultEffect, NameCardResultIntent>(
+    initialState = NameCardResultUiState()
+) {
+
+    init {
+        setUserInfo()
+    }
+
+    override fun processIntent(intent: NameCardResultIntent) {
+        when(intent) {
+            is NameCardResultIntent.Initialize -> {
+                updateState { copy(nameCardInfo = intent.nameCardInfo) }
+            }
+            is NameCardResultIntent.ShowShareBottomSheet -> {
+                updateState { copy(isShownShareBottomSheet = intent.isShown) }
+            }
+        }
+    }
+
+    private fun setUserInfo() = launch {
+        userRepository.getUser()?.let {
+            updateState { copy(user = UserUiData.fromDomain(it)) }
+        }
+    }
+}
